+++ src/notify/__init__.py
@@ -1,0 +1,3 @@
+"""
+src/notify/__init__.py - 通知模块
+"""